+++ src/features/App/types.ts
@@ -2,3 +2,9 @@ export type AppStatus = string|null
 export type Error = null|string
 export type Loading = 'idle'|'fulfilled'|'rejected'|'pending'
 
+export type DefaultPayload = {
+    dateFrom: string
+    dateTo: string
+    page: number
+    limit: number
+}
